# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ETMS API
  version: 1.0.0
  description: Employee Ticket Management System API
servers:
  - url: http://localhost:5003
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/tickets:
    post:
      tags: [Tickets]
      summary: Create ticket
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Ticket number conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Tickets]
      summary: List tickets
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending_approval, approved, assigned, in_progress, reported, resolved, closed, rejected]
        - in: query
          name: type
          schema:
            type: string
            enum: [complaint, request, data]
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high, critical]
        - in: query
          name: ticket_id
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: Ticket list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
...
  /api/assets:
    get:
      tags: [Assets]
      summary: List assets
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
    post:
      tags: [Assets]
      summary: Create asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
...
  /api/users:
    get:
      tags: [Users]
      summary: List users
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
...
  /api/data-portal/tickets:
    post:
      tags: [DataPortal]
      summary: Create data portal ticket
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateDataPortalTicketRequest'
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataPortalTicketCreateResponse'
    get:
      tags: [DataPortal]
      summary: List data portal tickets
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: Ticket list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        emp_id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [employee, manager, admin, data_team]
        team:
          type: string
          nullable: true
        category_id:
          type: integer
          nullable: true
        is_active:
          type: boolean
        can_manage_assets:
          type: boolean
          nullable: true
        password_changed_at:
          type: string
          nullable: true
    Ticket:
      type: object
      properties:
        id:
          type: integer
        ticket_no:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending_approval, approved, assigned, in_progress, reported, resolved, closed, rejected]
        priority:
          type: string
          enum: [low, medium, high, critical]
        type_key:
          type: string
          enum: [complaint, request, data]
        type_name:
          type: string
        category_id:
          type: integer
        category_name:
          type: string
        category_key:
          type: string
        raised_by:
          type: integer
        raised_by_name:
          type: string
        raised_by_emp_id:
          type: string
        raised_by_email:
          type: string
        assigned_to:
          type: integer
          nullable: true
        assigned_to_name:
          type: string
          nullable: true
        approval_owner_id:
          type: integer
          nullable: true
        asset_id:
          type: integer
          nullable: true
        asset_name:
          type: string
          nullable: true
        asset_serial_number:
          type: string
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        report_reason:
          type: string
          nullable: true
        sla_days:
          type: integer
        sla_due_date:
          type: string
          nullable: true
        escalated:
          type: boolean
          nullable: true
        escalated_at:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
...
    TicketListResponse:
      type: object
      properties:
        success:
          type: boolean
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
...
    TicketCreateResponse:
      type: object
      properties:
        success:
          type: boolean
        ticket:
          $ref: '#/components/schemas/Ticket'
...
    CreateTicketRequest:
      type: object
      required: [title, description, ticket_type_id, category_id, priority, sla_days]
      properties:
        title:
          type: string
        description:
          type: string
        ticket_type_id:
          type: integer
        category_id:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, critical]
        sla_days:
          type: integer
          minimum: 1
          maximum: 30
        asset_id:
          type: integer
        file:
          type: string
          format: binary
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        serial_number:
          type: string
        category_id:
          type: integer
        category_name:
          type: string
        assigned_to:
          type: integer
        assigned_user_name:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, under_repair, retired]
        machine_type:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        ram:
          type: string
          nullable: true
        hdd:
          type: string
          nullable: true
        monitor_serial:
          type: string
          nullable: true
        monitor_make:
          type: string
          nullable: true
        system_ip:
          type: string
          nullable: true
        port:
          type: string
          nullable: true
        ms_office_ver:
          type: string
          nullable: true
        os:
          type: string
          nullable: true
        host_id:
          type: string
          nullable: true
        floor:
          type: string
          nullable: true
        branch:
          type: string
          nullable: true
    AssetResponse:
      type: object
      properties:
        success:
          type: boolean
        asset:
          $ref: '#/components/schemas/Asset'
    AssetListResponse:
      type: object
      properties:
        success:
          type: boolean
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
...
    CreateAssetRequest:
      type: object
      required: [name, serial_number, category_id, assigned_to]
      properties:
        name:
          type: string
        serial_number:
          type: string
        category_id:
          type: integer
        assigned_to:
          type: integer
        status:
          type: string
          enum: [active, under_repair, retired]
        machine_type:
          type: string
        model:
          type: string
        ram:
          type: string
        hdd:
          type: string
        monitor_serial:
          type: string
        monitor_make:
          type: string
        system_ip:
          type: string
        port:
          type: string
        ms_office_ver:
          type: string
        os:
          type: string
        host_id:
          type: string
        floor:
          type: string
        branch:
          type: string
...
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    CreateUserRequest:
      type: object
      required: [emp_id, name, email, password, role]
      properties:
        emp_id:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
          enum: [employee, manager, admin, data_team]
        category_id:
          type: integer
          nullable: true
        team:
          type: string
          nullable: true
    CreateUserResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        auto_assigned_asset:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            name:
              type: string
            serial_number:
              type: string
...
    CreateDataPortalTicketRequest:
      type: object
      required: [title, description, ticket_type_id, category_id, priority, sla_days]
      properties:
        title:
          type: string
        description:
          type: string
        ticket_type_id:
          type: integer
        category_id:
          type: integer
        priority:
          type: string
        sla_days:
          type: integer
        file:
          type: string
          format: binary
    DataPortalTicketCreateResponse:
      type: object
      properties:
        success:
          type: boolean
        ticket:
          $ref: '#/components/schemas/Ticket'
        routed_team:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            category_key:
              type: string
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string